Refactor the association saver
In multiple area of the exporter we have to save the id of the shopify
object to the associated spree object.

The AssociationSaver does it for us.
